+++ composer.json
@@ -1,13 +1,13 @@
 {
     "name": "macocci7/php-boxplot",
-    "version": "1.2.0",
+    "version": "1.2.1",
     "description": "it's easy to use for creating boxplots.",
     "type": "library",
     "require": {
         "php": ">=8.1",
-        "intervention/image": "^3.5",
-        "macocci7/php-frequency-table": "1.3.0",
-        "fakerphp/faker": "^1.22",
+        "intervention/image": "^3.6",
+        "macocci7/php-frequency-table": "^1.3",
+        "fakerphp/faker": "^1.23",
         "nette/neon": "^3.4"
     },
     "require-dev": {
@@ -15,7 +15,7 @@
         "phpunit/phpunit": "^10.5",
         "phpmd/phpmd": "^2.15",
         "phpstan/phpstan": "^1.10",
-        "php-parallel-lint/php-parallel-lint": "^1.3"
+        "php-parallel-lint/php-parallel-lint": "^1.4"
     },
     "license": "MIT",
     "autoload": {
